Persist documents, one per author per file per commit
Each document has the author, the lines that author write for this file,
full filepath, commit SHA, timestamp

These documents are added to a Mongo database named "carry_your_weight",
into a collection named after the git project being analyed.

Currently, the database must be local, and the names above are not
configurable. This work must be brought in before bringing to the web.

Fixes #1