Wrap D3 logic in ActionCable
Previously, the index.html file acted as a mock server, a D3 client, and
used custom JS events to wire the two together.

This moves the (mock) server logic, parsing and transmitting a commit,
to the server, using ActionCable to push fake commits periodically. The
client-side logic handling these commits (in D3) is converted to use
ActionCable by copying the logic into the client-side "received" hook.

Currently, the ActionCable logic is depending HEAVILY on D3 setup logic
from the index.html page, which makes me uncomfortable, so I'll take a
look at getting as much of the logic as I can _out_ of index.html